updare one input file to all the layouts
